refactor: restructured project's architecture
