+++ src/presentation/components/button/BtnPlatform.tsx
