+++ src/presentation/lib/mocks/Platform.ts
